src/wallet-backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [{ name = "k688", email = "k688@k688.c" }]
requires-python = ">=3.13"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
]
dependencies = [
    "aiofiles>=25.1.0",
    "argon2-cffi>=25.1.0",
    "asyncpg>=0.31.0",
    "debugpy>=1.8.20",
    "quart (>=0.20.0,<0.21.0)",
    "quart-auth>=0.11.0",
    "sqlalchemy[asyncio]>=2.0.48",
]

[tool.uv]

[tool.uv.build-backend]
module-name = ["app"]
module-root = ""

[build-system]
requires = ["uv_build>=0.10.2,<0.11.0"]
build-backend = "uv_build"
